\name{convert.illions}
\alias{convert.illions}

\title{
Convert string to number
}
\description{
Convert a string such as "1.2M" to a number 1200000
}
\usage{
convert.illions(x)
}
\arguments{
  \item{x}{
string of numbers and a single letter e.g. "2.13B", "4T"
}
}
\details{
see description
}
\value{
numeric
}
\references{

}
\author{
Garrett See
}
\note{
}


\seealso{

}
\examples{

convert.illions("300M")
convert.illions("1.453T")
}
