\name{has.Mid}
\alias{has.Mid}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Does the data have a column.name containing \sQuote{Mid}
}
\description{
\code{has.Mid} checks to see if an xts object has a Mid column
}
\usage{
has.Mid(x, which = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
xts object
}
  \item{which}{
display position of match
}
}
\value{
%\code{Mid} returns a univariate xts object
\code{has.Mid} returns logical (boolean)
}
\author{
gsee
}
\note{
These should be exported from quantmod.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
Op, Hi, Lo, Cl, Bid, Ask,
has.Op, has.Cl
}

