\docType{package}
\name{qmao-package}
\alias{qmao}
\alias{"qmao-package"}
\alias{qmao-package}
\title{Some quantmod-like functions}
\description{
  more functions for quantitative trade modelling
}
\details{
  \tabular{ll}{ Package: \tab qmao\cr Type: \tab Package\cr
  Version: \tab 1.0\cr Date: \tab 2011-05-27\cr Depends:
  \tab quantmod(>=0.3-15),FinancialInstrument\cr License:
  \tab GPL-3\cr LazyLoad: \tab yes\cr } A few functions
  that are often useful when working in the xts/quantmod
  framework
}
\author{
  Garrett See

  Maintainer: Garrett See <gsee000@gmail.com>
}

