\name{SPDRSymbols}
\alias{SPDRSymbols}
\title{Get the ticker symbols of all SPDR ETFs}
\usage{
  SPDRSymbols()
}
\value{
  a character vector of the ticker symbols of all SPDR
  ETFs.
}
\description{
  Get a character vector of the ticker symbols of all SPDR
  ETFs.
}
\examples{
\dontrun{
SPDRSymbols()
}
}
\author{
  Garrett See
}
\references{
  \url{https://www.spdrs.com/product}
}
\seealso{
  getHoldings.SPDR
}

